requirements.txt
python-socketio
aiohttp
requests
schedule